pyproject.toml
[tool.poetry]
name = "resume-x"
version = "0.1.0"
description = ""
authors = ["Ben Vincent <benjc.vincent@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
pyyaml = "^6.0.2"
uritemplate = "^4.1.1"
django-oauth-toolkit = "^3.0.1"
drf-dynamic-fields = "^0.4.0"
django-cors-headers = "^4.6.0"
drf-access-policy = "^1.5.0"
pillow = "^11.0.0"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
drf-url-filters = "^0.5.1"
django-rest-polymorphic = "^0.1.10"
django-reversion = "^5.1.0"
django-polymorphic = {git = "https://github.com/jazzband/django-polymorphic.git", rev = "v4.0.0a"}
openai = "^1.66.3"
jinja2 = "^3.1.6"
python-jwt = "^4.1.0"
requests = "^2.32.3"
unique-names-generator = "^1.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
